requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装

-r requirements_services.txt
setuptools==57.5.0
django==3.2.4
blueapps==4.5.0
gunicorn==19.6.0
xlrd==0.9.4
xlwt==1.3.0
celery==4.4.5
Pillow==8.4.0
